requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dynaconf==3.2.11
hvac==2.3.0
idna==3.10
knox==0.1.14
loguru==0.7.3
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
shortuuid==1.0.13
sqlparse==0.5.3
urllib3==2.5.0
validators==0.35.0
